Add files via upload
In Java, this and super are keywords used to refer to objects within a class. They serve different purposes related to the current object and its superclass.this Keyword
The this keyword is a reference to the current object of the class. It is primarily used to: 
Access instance variables when there is a naming conflict with method parameters or local variables.
Call another constructor of the same class (constructor chaining).
Return the current object from a method.

super Keyword
The super keyword is a reference to the superclass (parent class) of the current class. It is used to: 
Access superclass members (fields and methods) that are hidden or overridden in the subclass.
Call a superclass constructor from a subclass constructor.