Add files via upload
A String in Java is an object that represents a sequence of characters. It's part of the java.lang package, and unlike arrays, it's immutable — which means once it's made, you can’t change it. (Like that friend who never changes their playlist.)There are so many methods that do different things, from comparison to transformation, and even the ability to check for patterns in text. Some of them are used in these program